+++ pyro/__init__.py
@@ -6,7 +6,6 @@ import logging
 import warnings
 from collections import OrderedDict
 from inspect import isclass
-import greenlet
 
 import torch
 from torch.autograd import Variable
@@ -14,9 +13,8 @@ from torch.autograd import Variable
 import pyro.poutine as poutine
 from pyro.distributions.distribution import Distribution
 from pyro.params import _MODULE_NAMESPACE_DIVIDER, _PYRO_PARAM_STORE, param_with_module_name
-from pyro.poutine import _PYRO_STACK, condition, do  # noqa: F401
+from pyro.poutine import _PYRO_STACK, condition, do, am_i_wrapped, send_message  # noqa: F401
 from pyro.util import apply_stack, deep_getattr, get_tensor_data, ones, set_rng_seed, zeros  # noqa: F401
-import pyro.util as util
 
 __version__ = '0.1.2'
 
@@ -57,7 +55,7 @@ def sample(name, fn, *args, **kwargs):
     baseline = kwargs.pop("baseline", {})
     # check if stack is empty
     # if stack empty, default behavior (defined here)
-    if not util.am_i_wrapped():
+    if not am_i_wrapped():
         if obs is not None:
             warnings.warn("trying to observe a value outside of inference at " + name,
                           RuntimeWarning)
@@ -87,7 +85,7 @@ def sample(name, fn, *args, **kwargs):
 
         # apply the stack and return its return value
         msg = apply_stack(msg)
-        out_msg = greenlet.getcurrent().parent.switch(msg)
+        out_msg = send_message(msg)
         return out_msg["value"]
 
         # output message and return the reply
@@ -237,7 +235,7 @@ def iarange(name, size=None, subsample_size=None, subsample=None, use_cuda=None)
     extended discussion.
     """
     subsample, scale = _subsample(name, size, subsample_size, subsample, use_cuda)
-    if not util.am_i_wrapped():
+    if not am_i_wrapped():
         yield subsample
     else:
         with poutine.scale(None, scale):
@@ -274,7 +272,7 @@ def irange(name, size, subsample_size=None, subsample=None, use_cuda=None):
     subsample, scale = _subsample(name, size, subsample_size, subsample, use_cuda)
     if isinstance(subsample, Variable):
         subsample = subsample.data
-    if not util.am_i_wrapped():
+    if not am_i_wrapped():
         for i in subsample:
             yield i
     else:
@@ -322,7 +320,7 @@ def param(name, *args, **kwargs):
     :param name: name of parameter
     :returns: parameter
     """
-    if not util.am_i_wrapped():
+    if not am_i_wrapped():
         return _PYRO_PARAM_STORE.get_param(name, *args, **kwargs)
     else:
         msg = {
@@ -337,7 +335,7 @@ def param(name, *args, **kwargs):
             "stop": False,
         }
         # apply the stack and return its return value
-        out_msg = greenlet.getcurrent().parent.switch(msg)
+        out_msg = send_message(msg)
         return out_msg["value"]
